requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
babel==2.17.0
    # via courlan
beautifulsoup4==4.13.4
    # via yfinance
certifi==2025.6.15
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
    #   trafilatura
cffi==1.17.1
    # via curl-cffi
charset-normalizer==3.4.2
    # via
    #   htmldate
    #   requests
    #   trafilatura
click==8.2.1
    # via uvicorn
coloredlogs==15.0.1
    # via onnxruntime
comm==0.2.2
    # via ipykernel
courlan==1.3.2
    # via trafilatura
curl-cffi==0.11.4
    # via yfinance
dateparser==1.2.1
    # via htmldate
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
executing==2.2.0
    # via stack-data
fastapi==0.115.13
    # via -r requirements.in
fastembed==0.7.1
    # via qdrant-client
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==25.2.10
    # via onnxruntime
frozendict==2.4.6
    # via yfinance
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
grpcio==1.73.0
    # via qdrant-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hdbscan==0.8.40
    # via -r requirements.in
hf-xet==1.1.5
    # via huggingface-hub
hpack==4.1.0
    # via h2
htmldate==1.9.3
    # via trafilatura
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   qdrant-client
huggingface-hub==0.33.0
    # via
    #   fastembed
    #   sentence-transformers
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipykernel==6.29.5
    # via -r requirements.in
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via
    #   hdbscan
    #   scikit-learn
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
justext==3.0.2
    # via trafilatura
loguru==0.7.3
    # via fastembed
lxml==5.4.0
    # via
    #   htmldate
    #   justext
    #   lxml-html-clean
    #   trafilatura
lxml-html-clean==0.4.2
    # via lxml
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mmh3==5.1.0
    # via fastembed
mpmath==1.3.0
    # via sympy
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.11
    # via yfinance
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   fastembed
    #   hdbscan
    #   onnxruntime
    #   pandas
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   transformers
    #   yfinance
onnxruntime==1.22.0
    # via fastembed
openai==1.91.0
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   ipykernel
    #   onnxruntime
    #   transformers
pandas==2.3.0
    # via yfinance
parso==0.8.4
    # via jedi
peewee==3.18.1
    # via yfinance
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via
    #   fastembed
    #   sentence-transformers
platformdirs==4.3.8
    # via
    #   jupyter-core
    #   yfinance
portalocker==2.10.1
    # via qdrant-client
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via
    #   onnxruntime
    #   qdrant-client
    #   yfinance
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py-rust-stemmers==0.1.5
    # via fastembed
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.0
    # via -r requirements.in
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   htmldate
    #   jupyter-client
    #   pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via
    #   dateparser
    #   pandas
    #   yfinance
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
qdrant-client==1.14.3
    # via -r requirements.in
regex==2024.11.6
    # via
    #   dateparser
    #   tiktoken
    #   transformers
requests==2.32.4
    # via
    #   fastembed
    #   huggingface-hub
    #   sec-api
    #   tiktoken
    #   transformers
    #   yfinance
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via
    #   hdbscan
    #   sentence-transformers
scipy==1.15.3
    # via
    #   hdbscan
    #   scikit-learn
    #   sentence-transformers
sec-api==1.0.32
    # via -r requirements.in
sentence-transformers==4.1.0
    # via -r requirements.in
setuptools==80.9.0
    # via torch
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via -r requirements.in
tld==0.13.1
    # via courlan
tokenizers==0.21.1
    # via
    #   fastembed
    #   transformers
torch==2.7.1
    # via sentence-transformers
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   fastembed
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
trafilatura==2.0.0
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.52.4
    # via sentence-transformers
typing-extensions==4.14.0
    # via
    #   beautifulsoup4
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via dateparser
urllib3==2.5.0
    # via
    #   courlan
    #   htmldate
    #   qdrant-client
    #   requests
    #   trafilatura
uvicorn==0.34.3
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via yfinance
yarl==1.20.1
    # via aiohttp
yfinance==0.2.63
    # via -r requirements.in
